move the metrics to calpit.metrics
